refactor: encapsultate ingestor utils
